update whit sweetalert

+++ Proyectos/06-crud-firebase/src/components/ShowProductTable.jsx
@@ -3,29 +3,44 @@ import { useEffect } from 'react'
 import { useState } from 'react'
 import { deleteProduct, getProducts } from '../firebase/productosApi'
 import { useNavigate } from 'react-router-dom'
-
-const ShowProductTable = ({add}) => {
+import  Swal  from 'sweetalert2'
+const ShowProductTable = ({ add }) => {
   const [loading, setLoading] = useState(true)
   const [products, setProducts] = useState([])
   const navigate = useNavigate()
 
   const fechDataProducts = async () => {
-    try{
+    try {
       const data = await getProducts()
       setProducts(data)
-    }catch(err){
-      console.log("Error al cargar los productos",err);
-    }finally{
+    } catch (err) {
+      console.log("Error al cargar los productos", err);
+    } finally {
       setLoading(false)
     }
   }
   const handleDeleteProduct = async (id) => {
-    // console.log(id)
-    try{
-      await deleteProduct(id)
-      setProducts(products.filter(product => product.id !== id))
-    }catch(err){
-
+    try {
+      const response = await Swal.fire({
+        icon: 'warning',
+        title: '¿Estas seguro?',
+        text: 'No podras revertir esto',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Si, borrar',
+        cancelButtonText: 'Cancelar',
+      })
+      if(response.isConfirmed){
+        await deleteProduct(id)
+        setProducts(products.filter(product => product.id !== id))
+      }
+    } catch (err) {
+      Swal.fire({
+        icon: 'error',
+        title: 'Oops...',
+        text: 'Algo salio mal',
+      })
     }
   }
 
@@ -35,10 +50,10 @@ const ShowProductTable = ({add}) => {
     console.log(id)
     navigate(`/productos/${id}`)
   }
-
+  // useEffect para cargar los productos siempre que se haga un setProductos y 
   useEffect(() => {
     fechDataProducts()
-  },[setProducts,add])
+  }, [setProducts, add])
 
 
   return (
@@ -59,22 +74,22 @@ const ShowProductTable = ({add}) => {
                 </tr>
               </thead>
               <tbody>
-               {
+                {
                   products.map((product) => (
                     <tr key={product.id}>
                       <td className='py-2 px-4 border-b'>{product.id}</td>
                       <td className='py-2 px-4 border-b'>{product.name}</td>
                       <td className='py-2 px-4 border-b'>{product.stock}</td>
                       <td className='py-2 px-4 border-b'>{product.description}</td>
-                      <td className='py-2 px-4 border-b'>{product.url}</td>
+                      <td className='py-2 px-4 border-b'><img src={product.url} alt={`imagen del producto ${product.name}`} /></td>
                       <td className='py-2 px-4 border-b'>
-                        <button className='bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded' onClick={() => handleUpdateProduct(product.id)}>Editar</button>
+                        <button className='bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded m-2' onClick={() => handleUpdateProduct(product.id)}>Editar</button>
                         <button className='bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded' onClick={() => handleDeleteProduct(product.id)}>Eliminar</button>
                       </td>
                     </tr>
                   ))
-               }
-           
+                }
+
               </tbody>
             </table>
